Add run_dir into Calculators and check aligned_alloc for aligned allocation

+++ PyBigDFT/BigDFT/Calculators.py
@@ -234,29 +234,38 @@ class SystemCalculator(Runner):
         #Create a temporary dictionary of options
         self._run_options(**kwargs)
         #Give the default values in case of unset_global_options are called
-        name=self.run_options.get('name','')
-        outdir=self.run_options.get('outdir','')
-        run_dir=self.run_options.get('run_dir','.')
+        name = self.run_options.get('name','')
+        outdir = self.run_options.get('outdir','')
+        run_dir = self.run_options.get('run_dir','.')
+        run_name = self.run_options.get('run_name','')
+        posinp = self.run_options.get('posinp',None)
+        inp = self.run_options.get('input',{})
+        verbose = self.run_options['verbose']
+        dry_run = self.run_options['dry_run']
         #Restrict run_dir to a sub-directory
-        assert not "/" in run_dir 
-        run_name=self.run_options.get('run_name','')
-        inp=self.run_options.get('input',{})
-        posinp=self.run_options.get('posinp',None)
-        #Create the run_dir
+        if  ("/" in run_dir or run_dir == ".."):
+            raise ValueError("run_dir '%s' where bigdft is executed must be a sub-directory" % run_dir)
+        #Create the run_dir if not exist
         if not os.path.exists(run_dir):
             #Creation of the sub-directory run_dir
             os.mkdir(run_dir)
-        #Create the input file
+            if verbose: safe_print("Create the sub-directory '%s'" % run_dir)
+        #Create the input file (deepcopy because we modify it!)
         local_input=copy.deepcopy(inp)
+        #Check posinp file
         if posinp != None:
             #Check if the file does exist
-            assert os.path.isfile(posinp)
+            if not os.path.isfile(posinp):
+                raise ValueError("posinp: The atomic position file '%s' does not exist" % posinp)
             #Add into the dictionary a posinp key
             if run_dir == '.':
                 local_input['posinp'] = posinp
             else:
-                #Copy the posinp
-                shutil.copy2(posinp,run_dir)
+                #Copy the posinp if not identical
+                cp_posinp = "%s/%s" % (run_dir,posinp)
+                if not (os.path.isfile(cp_posinp) and os.stat(posinp) != os.stat(cp_posinp)):
+                    shutil.copy2(posinp,run_dir)
+                    if verbose: safe_print("Copy the posinp file '%s' into '%s'" % (posinp,run_dir))
                 local_input['posinp'] = os.path.basename(posinp)
         if len(name) > 0:
             input_file = "%s/%s.yaml" % (run_dir,name)
@@ -269,13 +278,13 @@ class SystemCalculator(Runner):
         #Creating the yaml input file
         from futile import YamlIO as Y
         Y.dump(local_input,filename=input_file)
-        if self.run_options['verbose']: safe_print('Creating the yaml input file "%s"' % input_file)
+        if verbose: safe_print('Creating the yaml input file "%s"' % input_file)
         #Check if it is a dry run
-        if self.run_options['dry_run']:
+        if dry_run:
             #Use bigdft-tool
             command = self.run_options['mpi_run'] + ' ' + os.environ['BIGDFT_ROOT']+'/bigdft-tool -l'
-            if self.run_options['dry_run']:
-                command += ' -n ' + str(self.run_options['dry_run'])
+            if dry_run:
+                command += ' -n ' + str(dry_run)
             if len(name) > 0:
                 command += ' --name='+name
         else:   
@@ -291,14 +300,14 @@ class SystemCalculator(Runner):
                 logname = outdir+"/"+logname
             if self.run_options['skip']:
                 command += ' -s Yes'
-        if self.run_options['verbose']: safe_print('Executing command: ', command)
+        if verbose: safe_print('Executing command: ', command)
         # Set the number of omp threads
         os.environ['OMP_NUM_THREADS'] = self.run_options['omp']
         #Run the command
         os.system("cd "+run_dir+"; "+command)
         #verify that no debug file has been created
         if get_debugfile_date() > timedbg :
-            if self.run_options['verbose']: 
+            if verbose: 
                 safe_print("ERROR: some problem occured during the execution of the command, check the 'debug/' directory and the logfile")
             return None
         return self.post_processing(logname)
